requirements.txt
# HTTP y scraping
requests==2.28.1
beautifulsoup4==4.11.1
selenium==4.22.0
webdriver-manager==4.0.1

# Manejo de datos
numpy==1.26.4
pandas==2.1.2 

# Google Cloud
google-cloud-storage==2.7.0

# Base de datos
SQLAlchemy==1.4.46
PyMySQL==1.0.2

# Utilidades
python-dotenv==1.1.1
cryptography==41.0.3
geopy==2.4.1
pendulum==3.1.0

# Airflow y proveedores compatibles
apache-airflow==2.8.4
apache-airflow-providers-mysql==5.7.4
apache-airflow-providers-google==11.0.0
astronomer-cosmos==1.2.0

#Dbt
dbt-mysql==1.7.0

#Visualizaciones
matplotlib==3.8.0
seaborn==0.12.2
scipy==1.11.2